move getConfig, getExtensionsInfo into utils.js
