+++ js/popup.js
@@ -1,6 +1,6 @@
 import { Component, h, render } from './vendor/preact-10.0.1.js'
 import htm from './vendor/htm-2.2.1.js'
-import { getConfig, getExtensionsInfo } from './background.js'
+import { getConfig, getExtensionsInfo } from './utils.js'
 
 const currentVersion = window.navigator.userAgent.match(
   /Chrome\/([0-9]+\.[0-9]+\.[0-9]+\.[0-9]+)/
